2023.11.10. 업데이트. 상품 상세페이지 기초 작성. 헤더의 장바구니 갱신 및 마지막 상품 삭제 안되는거 수정필요성 확인

+++ Project/src/main/java/com/project/controller/kjs/ShoppingCartController.java
@@ -1,6 +1,5 @@
 package com.project.controller.kjs;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.sql.SQLException;
 import java.util.HashMap;
@@ -17,13 +16,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.util.WebUtils;
@@ -31,7 +27,7 @@ import org.springframework.web.util.WebUtils;
 import com.project.controller.HomeController;
 import com.project.service.kjs.shoppingcart.ShoppingCartService;
 import com.project.vodto.Member;
-import com.project.vodto.Product;
+import com.project.vodto.kjs.DisPlayedProductDTO;
 
 /**
  * @author goott1
@@ -57,16 +53,6 @@ public class ShoppingCartController {
 		return mav;
 	}
 	
-	
-	/**
-	 * @MethodName : shoppingCartList
-	 * @author : goott1
-	 * @param model
-	 * @param request
-	 * @returnType : void
-	 * @description : 장바구니 컨트롤러 - 내역 불러오기
-	 * @date : 2023. 10. 13.
-	 */
 	@RequestMapping("shoppingCart/all")
 	public ResponseEntity<Map<String, Object>> shoppingCartList(HttpServletRequest request) {
 		System.out.println("======= 장바구니 컨트롤러 - 장바구니 내역 불러오기 =======");
@@ -78,27 +64,16 @@ public class ShoppingCartController {
 		try {
 			if (session.getAttribute("loginMember") != null) {
 				// 로그인을 했으면
-				list = scService.getShoppingCart(((Member)session.getAttribute("loginMember")).getMemberId(), true);
-				List<Product> items = (List<Product>)list.get("items");
+				Member member = (Member) session.getAttribute("loginMember");
+				list = scService.getShoppingCart(member.getMemberId(), true);
+				List<DisPlayedProductDTO> items = (List<DisPlayedProductDTO>)list.get("items");
 				result = output(list, items);
 			} else {
 				// 로그인을 안했으면(비회원이면)			
 				Cookie cookie = WebUtils.getCookie(request, "nom");
 				if (cookie != null) {
 					list = scService.getShoppingCart(cookie.getValue(), false);
-					// 해당 비회원의 상품 장바구니 List
-//					List<ShoppingCart> cartList = (List<ShoppingCart>)list.get("list");
-					// 장바구니 List에 담긴 상품의 정보
-					List<Product> items = (List<Product>)list.get("items");
-//					model.addAttribute("cartList", cartList);
-//					if (items != null && items.size() > 0) {
-//						map.put("items", items);
-//						map.put("status", "success");
-//					}
-//					else {
-//						map.put("status", "none");
-//					}
-//					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
+					List<DisPlayedProductDTO> items = (List<DisPlayedProductDTO>)list.get("items");
 					result = output(list, items);
 				} // else : 타이밍 맞게 쿠키 유효기간이 지나 삭제됐을 경우
 			}			
@@ -111,7 +86,7 @@ public class ShoppingCartController {
 		return result;
 	}
 	
-	private ResponseEntity<Map<String,Object>> output(Map<String, Object> list, List<Product> items) {
+	private ResponseEntity<Map<String,Object>> output(Map<String, Object> list, List<DisPlayedProductDTO> items) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		
 		if (items != null && items.size() > 0) {
@@ -125,42 +100,40 @@ public class ShoppingCartController {
 		return new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 	}
 	
-	/**
-	 * @MethodName : deleteItem
-	 * @author : goott1
-	 * @param request
-	 * @param productId
-	 * @return
-	 * @returnType : ResponseEntity<Map<String,Object>>
-	 * @description : 장바구니 컨트롤러 - 아이템(1개) 삭제
-	 * @date : 2023. 10. 13.
-	 */
 	@RequestMapping(value="{productId}", method=RequestMethod.DELETE)
 	public ResponseEntity<Map<String, Object>> deleteItem(HttpServletRequest request, @PathVariable("productId") String productId) {
 		System.out.println("======= 장바구니 컨트롤러 - 아이템(1개) 삭제 =======");
 		ResponseEntity<Map<String, Object>> result = null;
 		Map<String, Object> map = new HashMap<String, Object>();
 		// 로그인 여부 확인
-		// 로그인을 했으면
+		HttpSession session = request.getSession();
 		
-		// 비회원이면
-		Cookie cookie = WebUtils.getCookie(request, "nom");
-		System.out.println("id : " + cookie.getValue() + ", productId : " + productId);
-		if (cookie != null) {
-			try {				
-				// 하나 이상의 row가 삭제되었다면
-				if (scService.deleteItem(cookie.getValue(), false, productId)) {
+		try {				
+			if (session.getAttribute("loginMember") != null) {
+				// 로그인을 했으면
+				if (scService.deleteItem(((Member)session.getAttribute("loginMember")).getMemberId(), true, productId)) {
 					map.put("status", "success");
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 				}
-				else {
-					map.put("status", "fail");
+			} else {
+				// 비회원이면
+				Cookie cookie = WebUtils.getCookie(request, "nom");
+				System.out.println("id : " + cookie.getValue() + ", productId : " + productId);
+				if (cookie != null) {
+					// 하나 이상의 row가 삭제되었다면
+					if (scService.deleteItem(cookie.getValue(), false, productId)) {
+						map.put("status", "success");
+					}
+					else {
+						map.put("status", "fail");
+					}
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 				}
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
-			} catch (SQLException | NamingException e) {
-				map.put("status", "fail");
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
-				e.printStackTrace();
 			}
+		} catch (SQLException | NamingException e) {
+			map.put("status", "fail");
+			result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
+			e.printStackTrace();
 		}
 		
 		System.out.println("======= 장바구니 컨트롤러단 끝 =======");
@@ -201,26 +174,41 @@ public class ShoppingCartController {
 	}
 	
 	// item 추가
-	@RequestMapping(value="{productId}", method=RequestMethod.POST)
-	public ResponseEntity<Map<String, Object>> addProduct(HttpServletRequest request, @PathVariable("productId") String productId) {
+	@RequestMapping("insert")
+	public ResponseEntity<Map<String, Object>> addProduct(HttpServletRequest request, @RequestParam("product_id") String productId) {
 		System.out.println("======= 장바구니 컨트롤러 - 아이템 추가 =======");
 		ResponseEntity<Map<String, Object>> result = null;
 		Map<String,Object> map = new HashMap<String, Object>();
-		// 로그인 여부 확인
-		// 로그인 했을 시
 		
-		// 비회원일 시
-		Cookie cookie = WebUtils.getCookie(request, "nom");
-		if (cookie != null) {
-			try {
-				scService.insertItem(cookie.getValue(), false, productId);
-				map.put("status", "success");
+		// 로그인 여부 확인
+		HttpSession session = request.getSession();
+		try {
+			if (session.getAttribute("loginMember") != null) {
+				// 로그인 했을 시
+				if (scService.insertItem(((Member)session.getAttribute("loginMember")).getMemberId(), true, productId)) {
+					map.put("status", "success");
+				} else {
+					// 이미 담긴 상품일 시
+					map.put("status", "exist");
+				}
 				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
-			} catch (SQLException | NamingException e) {
-				e.printStackTrace();
-				map.put("status", "fail");
-				result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
+			} else {
+				// 비회원일 시
+				Cookie cookie = WebUtils.getCookie(request, "nom");
+				if (cookie != null) {
+					if (scService.insertItem(cookie.getValue(), false, productId)) {
+						map.put("status", "success");
+					} else {
+						// 이미 담긴 상품일 시
+						map.put("status", "exist");
+					}
+					result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
+				}
 			}
+		} catch (SQLException | NamingException e) {
+			e.printStackTrace();
+			map.put("status", "fail");
+			result = new ResponseEntity<Map<String,Object>>(map, HttpStatus.BAD_REQUEST);
 		}
 		
 //		String history = request.getHeader("referer");
@@ -232,9 +220,12 @@ public class ShoppingCartController {
 		return result;
 	}
 	
-	@RequestMapping("/insert")
-	public ResponseEntity<String> test(@RequestParam("product_id") String id){
-		System.out.println(id + "장바구니 들어옴");
-		return ResponseEntity.ok("Success");
+	// 결제페이지에 데이터 넘기기
+	public void orderPage(HttpServletRequest request) {
+		System.out.println("======= 장바구니 컨트롤러 - 결제페이지 이동 =======");
+		
+		
+		
+		System.out.println("======= 장바구니 컨트롤러단 끝 =======");
 	}
 }
